Renamed modules; made phase spliting typesafe.
Phasesplit is now know and phasealign.
Streamphase is now known as phasesplit.

Closes #35.

We reintroduce phases; these are typesafe abstractions. Global
declarations have a single element in their body; the unified
instructions of before removed this restriction. Furthermore, in the
serialization, because global-code only has one instruction in each
block, we can linearize it and make it easier to read/parse.