Simplifying logic of steps in ranges.
Reified the logic of predicates in the module Range, which holds the
code for loop ranges. Multiplicative ranges are now handled in a type
safe way. There are no extension points.

In range, added better support for generating the last iteration with
multiplicative loops with symbolic steps.

Moved the pretty-printing logic from Serialize into each data-structure.
Deleted serialize as this is mostly legacy code needed before for
smtlib2 serialization.