Revised alarm output.
Firstly, alarms were showing more variables than needed.
This was due to the result from Z3 includes variables from
all concurrent accesses, yet when we report we only want to
show the information relevant to the 2 racy accesses, not
the state from all concurrent variables (which are unrelated). We now
filter the output of Z3 to only show
variables from the 2 racy accesses.

Secondly, we revised the messages reported regarding
approximation analysis. We summarize the status, say CIDI,
in the title of the data-race and we annotate with (C) or (D) each
variable. Additionally, we clearly flag when the alarm is true.
We have more consistent warnings for CD/DD (before we were only showing
data-approx), potential alarm warning, reason (control or data
dependence) and explain root cause (look at variables).