Add better support for calling kernels with generics.
Calls to kernels with generics may not have a type signature available.
In which case, our data-base now needs to fallback into loading an
arbitraty kernel with the same name. The potential issue is if there are
multiple kernels we may not pick the right one, but I cannot think of a
better solution when there is no typing information available.