Fixed a bug in the bit-vector backend.
With the addition of signed information, the bit-vector backend became
unsound because the bit-vector representation (32bits) was too small to
represent signed and unsigned integers, since these are 32bits in width.
We now use 64bits and assume all numbers are signed.

We also added an initial unreachable check to find unreachable accesses
(which allowed us to confirm this bug).