Improved the accuracy of how asserts are encoded.
The pattern if (cond) return; was being recognized as an assert.
However, this kind of assert cannot be hoisted, whereas as "standard"
assert should be hoisted to the top of the kernel (as in, its effects
are global). We now have a distinction between local (such as if cond
return) and global (the actual asserts). This means that our tool can
detect more CIDI data-races than before.