Simplified the inference of WGSL to reuse Infer_stmt.
w-to-imp no longer needs to infer Exp and uses a much simpler loop
inference (we simplify convert loop to a for-like loop). With the former
change the monad state is much simpler, as we only have to keep track of
reads. Additionally, with this change we now keep track of the
expression being returned.