Added support for handling unif * tid.
In developing this, there was a need to introduce Clamp which produces
a tick of an expression that is clamped to an upper bound. This had to
be implemented as a primitive data-type to allow Maxima to produce
better code, as otherwise we would be observing verbose expressions
(as some summations were not being simplified).